# various optimizations