@User4martin 2023-11-20
Yes.

Only `(?:` was changed.

A normal `(` must still omit an opcode => so the capture can be saved (and it must support backtracking (in 99% of cases) so it needs recursion).

`(?>` still produces OP_OPEN_ATOMIC.  And that is also still needed (Though there is a limited subset of cases, where it could be optimized...)
- Needed: `a(?>bb|b)b` on text `abb` => will NOT match, as the atomic eats both `b`, and the b after the bracket has nothing to match.
- Subject to future optimization: `a(?>bb|b){1,3}b` . In this very specific case the atomic could (afaik) be provided by the loop `a(?:bb|b){1,3}+b`. This is only the case if the atomic bracket is exactly the entire loop.