@Alexey-T 2023-11-21
>A normal ( must still omit an opcode

Omit VS Emit. you mean 'emit'.